param tuning unit test
