+++ tests/orbit/models/test_lgt.py
@@ -5,7 +5,6 @@ import numpy as np
 from orbit.models import LGT
 from orbit.template.lgt import LGTInitializer
 from orbit.constants.constants import PredictionKeys
-from orbit.utils.params_tuning import grid_search_orbit
 
 
 @pytest.mark.parametrize("estimator", ["stan-map", "stan-mcmc"])
@@ -530,47 +529,6 @@ def test_lgt_fixed_sm_input(
     assert predict_df.columns.tolist() == expected_columns
     assert regression_out.shape == expected_regression_shape
     assert num_regressors == len(train_df.columns.tolist()[2:])
-
-
-@pytest.mark.parametrize(
-    "param_grid",
-    [
-        {
-            "level_sm_input": [0.3, 0.5, 0.8],
-            "seasonality_sm_input": [0.3, 0.5, 0.8],
-        },
-        {
-            "level_sm_input": [0.3, 0.5, 0.8],
-            "slope_sm_input": [0.3, 0.5, 0.8],
-        },
-    ],
-)
-def test_lgt_grid_tuning(make_weekly_data, param_grid):
-    train_df, test_df, coef = make_weekly_data
-    args = {
-        "response_col": "response",
-        "date_col": "week",
-        "seasonality": 52,
-        "estimator": "stan-map",
-    }
-
-    lgt = LGT(**args)
-
-    best_params, tuned_df = grid_search_orbit(
-        param_grid,
-        model=lgt,
-        df=train_df,
-        min_train_len=80,
-        incremental_len=20,
-        forecast_len=20,
-        metrics=None,
-        criteria=None,
-        verbose=True,
-    )
-
-    assert best_params[0].keys() == param_grid.keys()
-    assert set(tuned_df.columns.to_list()) == set(list(param_grid.keys()) + ["metrics"])
-    assert tuned_df.shape == (9, 3)
 
 
 @pytest.mark.parametrize("estimator", ["stan-mcmc", "stan-map"])
